feat: Add a REST API rust backend based on Axum and OpenAPI
Use Axum to serve a REST API based on the statistics collected
in the database.
Use http-tower::ServeDir to serve static assets (the future web
app) in the debug profile.
Use rust_embed to serve the assets in release profile.
Use Aide to generate an OpenAPI schema based on the Axum routes.
This schema will then be used in the web app to parse and validate the
REST API types. This allows declaring types in Rust, and then using
them in typescript using the OpenAPI schema as the in-between
converter.