refactor(config): Use std::time::Duration for update intervals
Replace usage of humantime::Duration with std::time::Duration in
UpdateInterval.

As a drive-by for the future, add serializer support to UpdateInterval.
Use humantime-serde crate for that.

Remove old unused fork that was in humantime_wrapper_serde.rs, there's
no need to store a humantime::Duration in the interval, it's
sufficient to store just an std::time::Duration and do parsing and
formatting as appropriate.

Presumably I wrote the fork because I thought I needed the FromStr
impl that humantime::Duration provided.