chore: Improve release process
Make sure to use the --latest flag for normal releases and
--unreleased --bump for test releases.

Use a separate tag name for test releases to avoid conflicts with the
branch name.

Modify the mise task for generating the changelog to generate a
changelog based on all commits.

Strip the git cliff header when generating changelogs for the github
release. It's only useful for a CHANGELOG.md in the actual repo.

Ensure we include unconventional commits in the repo changelog.

Make sure git cliff --bump always bumps the patch version for now, by
setting the cliff.toml breaking_always_bump_major option to false.