feat(backend): Add Hetzner Cloud DNS support
Added support for Hetzner Cloud DNS as a DNS provider.

There is no support for the old non-cloud DNS API. It is being phased
out, and it's not possible to test it with a new domain, so only the
new Cloud DNS API is supported.

Providers should be added via the new [[providers]] configuration
option in the config file.

See the updated sample config file for details.

Temporarily remove calls to init_color_eyre, because it causes
panics in multi-threaded tests due to race conditions on installing
the handler manually, and using eyre's undocumented auto-install
feature.

Fixes: #101